requirements.txt
psycopg2-binary
django
django-extensions
django-compressor
django-sass-processor
django-livereload-server
libsass
progress
pyproj
pyyaml
pause
pyinaturalist
wikipedia